Rename API endpoints
