+++ forge/ee/routes/teamBroker/index.js
@@ -346,284 +346,4 @@ module.exports = async function (app) {
         const list = await app.teamBroker.getUsedTopics(request.team.hashid)
         reply.send(list)
     })
-
-    /**
-     * Get All Credentials for 3rd Party Brokers linked to a team
-     * @name /api/v1/teams:/teamId/broker/credentials
-     * @static
-     * @memberof forge.routes.api.team.broker
-     */
-    app.get('/credentials', {
-        preHandler: app.needsPermission('broker:credentials:list'),
-        schema: {
-            summary: 'Get credentials for 3rd party MQTT brokers',
-            tags: ['MQTT Broker'],
-            params: {
-                type: 'object',
-                properties: {
-                    teamId: { type: 'string' }
-                }
-            },
-            response: {
-                200: {
-                    type: 'object',
-                    properties: {
-
-                    },
-                    additionalProperties: true
-                },
-                '4xx': {
-                    $ref: 'APIError'
-                },
-                500: {
-                    $ref: 'APIError'
-                }
-            }
-        }
-    }, async (request, reply) => {
-        const paginationOptions = app.getPaginationOptions(request)
-        const credsList = await app.db.models.BrokerCredentials.byTeam(request.params.teamId, paginationOptions)
-        reply.send(app.db.views.BrokerCredentials.cleanList(credsList))
-    })
-
-    /**
-     * Add new Credentials for a 3rd Party Broker
-     * @name /api/v1/teams/:teamId/broker/credentials
-     * @static
-     * @memberof forge.routes.api.team.broker
-     */
-    app.post('/credentials', {
-        preHandler: app.needsPermission('broker:credentials:create'),
-        schema: {
-            summary: 'Create credentials for a 3rd party MQTT broker',
-            tags: ['MQTT Broker'],
-            params: {
-                type: 'object',
-                properties: {
-                    teamId: { type: 'string' }
-                }
-            },
-            body: {
-                type: 'object',
-                properties: {
-                    name: { type: 'string' },
-                    host: { type: 'string' },
-                    port: { type: 'number' },
-                    protocol: { type: 'string' },
-                    protocolVersion: { type: 'number' },
-                    ssl: { type: 'boolean' },
-                    verifySSL: { type: 'boolean' },
-                    clientId: { type: 'string' },
-                    credentials: {
-                        type: 'object'
-                    }
-                }
-            },
-            response: {
-                201: {
-                    type: 'object',
-                    properties: {
-
-                    },
-                    additionalProperties: true
-                },
-                '4xx': {
-                    $ref: 'APIError'
-                },
-                500: {
-                    $ref: 'APIError'
-                }
-            }
-        }
-    }, async (request, reply) => {
-        // Need to create a Access Token then pass it to the container driver
-        // to spin up a mqtt-schema-agent
-        const input = request.body
-        input.state = 'stopped'
-        input.credentials = JSON.stringify(request.body.credentials)
-        input.TeamId = app.db.models.Team.decodeHashid(request.params.teamId)
-        const creds = await app.db.models.BrokerCredentials.create(input)
-        creds.refreshAuthTokens()
-        const clean = app.db.views.BrokerCredentials.clean(creds)
-        reply.status(201).send(clean)
-    })
-
-    /**
-     * Get Credentials for a 3rd Party Broker
-     * @name /api/v1/teams/:teamId/broker/:brokerId/credentials
-     * @static
-     * @memberof forge.routes.api.team.broker
-     */
-    app.get('/:brokerId/credentials', {
-        preHandler: [
-            async (request, reply) => {
-                // TODO Needs custom preHandler to work with token for mqtt agent only
-                if (request.session?.scope?.includes('broker:credentials')) {
-                    if (request.session.ownerType === 'broker') {
-                        if (request.params.teamId !== request.session.Broker.Team.hashid) {
-                            reply.code('401').send({ code: 'unauthorized', error: 'unauthorized' })
-                        }
-                    } else {
-                        reply.code('401').send({ code: 'unauthorized', error: 'unauthorized' })
-                    }
-                } else {
-                    reply.code('401').send({ code: 'unauthorized', error: 'unauthorized' })
-                }
-            }
-        ],
-        schema: {
-            summary: 'Gets credentials for a 3rd party MQTT broker',
-            tags: ['MQTT Broker'],
-            params: {
-                type: 'object',
-                properties: {
-                    teamId: { type: 'string' },
-                    brokerId: { type: 'string' }
-                }
-            },
-            response: {
-                200: {
-                    type: 'object',
-                    properties: {
-
-                    },
-                    additionalProperties: true
-                },
-                '4xx': {
-                    $ref: 'APIError'
-                },
-                500: {
-                    $ref: 'APIError'
-                }
-            }
-        }
-    }, async (request, reply) => {
-        // console.log(request.params)
-        const creds = await app.db.models.BrokerCredentials.byId(request.params.brokerId)
-        if (creds) {
-            if (creds.TeamId === request.params.teamId) {
-                const resp = creds.toJSON()
-                resp.id = resp.hashid
-                delete resp.hashid
-                delete resp.slug
-                delete resp.links
-                resp.credentials = JSON.parse(resp.credentials)
-                reply.send(resp)
-            } else {
-                reply.code('401').send({ code: 'unauthorized', error: 'unauthorized' })
-            }
-        } else {
-            reply.status(404).send({ code: 'not_found', error: 'not found' })
-        }
-    })
-
-    /**
-     * Edit 3rd Party Broker credentials
-     * @name /api/v1/teams/:teamId/broker/:brokerId
-     * @static
-     * @memberof forge.routes.api.team.broker
-     */
-    app.put('/:brokerId', {
-        preHandler: app.needsPermission('broker:credentials:edit'),
-        schema: {
-            summary: 'Delete credentials for a 3rd party MQTT broker',
-            tags: ['MQTT Broker'],
-            params: {
-                type: 'object',
-                properties: {
-                    teamId: { type: 'string' },
-                    brokerId: { type: 'string' }
-                }
-            },
-            body: {
-                type: 'object',
-                properties: {
-                    name: { type: 'string' },
-                    host: { type: 'string' },
-                    port: { type: 'number' },
-                    protocol: { type: 'string' },
-                    protocolVersion: { type: 'number' },
-                    ssl: { type: 'boolean' },
-                    verifySSL: { type: 'boolean' },
-                    clientId: { type: 'string' },
-                    credentials: {
-                        type: 'object'
-                    }
-                }
-            },
-            response: {
-                200: {
-                    type: 'object',
-                    properties: {
-
-                    },
-                    additionalProperties: true
-                },
-                '4xx': {
-                    $ref: 'APIError'
-                },
-                500: {
-                    $ref: 'APIError'
-                }
-            }
-        }
-    }, async (request, reply) => {
-        const brokerCreds = await app.db.models.BrokerCredentials.byId(request.params.brokerId)
-        if (request.body.credentials) {
-            request.body.credentials = JSON.stringify(request.body.credentials)
-        }
-        await brokerCreds.update(request.body)
-        const clean = app.db.views.BrokerCredentials.clean(brokerCreds)
-        reply.send(clean)
-    })
-
-    /**
-     * Remove 3rd Party Broker credentials
-     * @name /api/v1/teams/:teamId/broker/:brokerId
-     * @static
-     * @memberof forge.routes.api.team.broker
-     */
-    app.delete('/:brokerId', {
-        preHandler: app.needsPermission('broker:credentials:delete'),
-        schema: {
-            summary: 'Delete credentials for a 3rd party MQTT broker',
-            tags: ['MQTT Broker'],
-            params: {
-                type: 'object',
-                properties: {
-                    teamId: { type: 'string' },
-                    brokerId: { type: 'string' }
-                }
-            },
-            response: {
-                200: {
-                    type: 'object',
-                    properties: {
-
-                    },
-                    additionalProperties: true
-                },
-                '4xx': {
-                    $ref: 'APIError'
-                },
-                500: {
-                    $ref: 'APIError'
-                }
-            }
-        }
-    }, async (request, reply) => {
-        const creds = await app.db.models.BrokerCredentials.byId(request.params.brokerId)
-        if (creds) {
-            try {
-                // TODO Need to tear down the running mqtt-schema-agent
-                // and remove the AccessToken
-                await creds.destroy()
-                reply.send({})
-            } catch (err) {
-                reply.status(500).send({ error: 'unknown_erorr', message: err.toString() })
-            }
-        } else {
-            reply.status(404).send({})
-        }
-    })
 }
